fix#12 modify shell script code.

+++ setup.sh
@@ -1,9 +1,22 @@
 echo -e "\033[0;34mMade by \033[1;36mKeonHeeLee] \033[1;34m(Thursday, 7 June 2018)"
 echo -e "\033[0;34m- Github : \033[1;36mhttps://github.com/KeonHeeLee/study-compiler"
-echo -e "\033[0;34m- E-mail : \033[1;36mbeta1360@naver.com\e[39m"
+echo -e "\033[0;34m- E-mail : \033[1;36mbeta1360@naver.com\033[0m"
 echo " "
 
+if ! [ -d "$icg"]; then
+    echo -e "\033[0;31mSetup Error: Don't exist directory::icg\033[0m"
+    echo -e "\033[0;31mPlease check directory::icg\033[0m"
+    exit
+fi
+
 cd icg
+
+if ! [ -e "$make"]; then
+    echo -e "\033[0;31mSetup Error: Don't compile this source codes!!\033[0m"
+    echo -e "\033[0;31mMakefile don't exists this directory.\033[0m"
+    exit
+fi
+
 make
 mv minic ..
 make clean
